[#ic-31] remove previous version of UpsertedProfileOrchestrator
