+++ UpsertedProfileOrchestrator/index.ts
@@ -1,24 +1,11 @@
 ﻿import * as df from "durable-functions";
-import { pipe } from "fp-ts/lib/function";
-import * as NonEmptyArray from "fp-ts/lib/NonEmptyArray";
-import * as O from "fp-ts/lib/Option";
-
 import { getConfigOrThrow } from "../utils/config";
-
 import { getUpsertedProfileOrchestratorHandler } from "./handler";
 
 const config = getConfigOrThrow();
 
 const orchestrator = df.orchestrator(
   getUpsertedProfileOrchestratorHandler({
-    notifyOn: config.FF_NEW_USERS_EUCOVIDCERT_ENABLED
-      ? pipe(
-          NonEmptyArray.fromArray([
-            config.EUCOVIDCERT_PROFILE_CREATED_QUEUE_NAME
-          ]),
-          O.getOrElseW(() => undefined)
-        )
-      : undefined,
     sendCashbackMessage: config.IS_CASHBACK_ENABLED
   })
 );
